backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gitpython==3.1.40
tree-sitter==0.20.4
langchain==0.1.0
chromadb==0.4.18
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
pygithub==2.1.1
ollama==0.1.0
